added pseudo code to GameLogic
